+++ extension/CHANGELOG.md
@@ -37,8 +37,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
           - `<code lang="json">{code: "theCode"}</code>`
           - `<code language="al">Message('Hello World!');</code>`
       - If no language attribute is used, `al` will be used as the default language.
-        - To enable `al` in highlight.js, <https://github.com/microsoft/AL/blob/master/highlightjs_al/dist/al.min.js> can be used.
+        - To enable `al` in [highlight.js](https://highlightjs.org/), <https://github.com/microsoft/AL/blob/master/highlightjs_al/dist/al.min.js> can be used.
       - The language attribute is ignored for inline code.
+    - Fields and Sub Pages are printed on API pages and WebServices pages.
+    - The first line of the XmlComment `Summary` is now printed for table fields and page controls (Fields, Actions, Parts).
 - Changes:
   - A change has been made to the generated External Documentation (created from the `NAB: Generate External Documentation`):
     - The procedure signatures now gets the default language `al`. See above for info on highlight.js.